Compute periodogram from light curve in json format

+++ src/periodogram/information_theory.rs
@@ -1,7 +1,7 @@
 use nalgebra::{DMatrix, linalg};
 
-pub fn renyi_gram_entropy(gram_matrix: DMatrix<f64>, n: f64, alpha: f64) -> f64 {
-    let mut s = 0.0_f64;
+pub fn renyi_gram_entropy(gram_matrix: DMatrix<f32>, n: f32, alpha: f32) -> f32 {
+    let mut s = 0.0_f32;
     let eigenvalues = linalg::SymmetricEigen::new(gram_matrix/n).eigenvalues;
     for i in 0usize..eigenvalues.len() {
         let eigval = eigenvalues[i];
@@ -12,7 +12,7 @@ pub fn renyi_gram_entropy(gram_matrix: DMatrix<f64>, n: f64, alpha: f64) -> f64 
     return s.log2()/(1.0-alpha);    
 }
 
-pub fn renyi_gram_joint_entropy(a: DMatrix<f64>, b: DMatrix<f64>, n: f64, alpha: f64) -> f64 {
+pub fn renyi_gram_joint_entropy(a: DMatrix<f32>, b: DMatrix<f32>, n: f32, alpha: f32) -> f32 {
     
     return renyi_gram_entropy(a.component_mul(&b), n, alpha);
 }
